+++ app/Views/sitemap/index.php
@@ -3,7 +3,7 @@
 <?php if (is_array($query ?? '') && $query !== []) {
     foreach ($query as $row) { ?>
     <sitemap>
-        <loc><?= $_ENV['app_baseLocalGeohistoryProjectUrl'] ?? '' ?>/<?= \Config\Services::request()->getLocale() ?>/<?= ($row === 'other' ? '' : $row . '/') ?>sitemap.txt</loc>
+        <loc><?= $_ENV['app_baseCanonicalProjectUrl'] ?? '' ?>/<?= \Config\Services::request()->getLocale() ?>/<?= ($row === 'other' ? '' : $row . '/') ?>sitemap.txt</loc>
     </sitemap>
 <?php }
     } ?>