+++ app/Views/sitemap/view_other.php
@@ -1,6 +1,6 @@
 <?php if (is_array($query ?? '') && $query !== []) {
     foreach ($query as $row) { ?>
-<?= $_ENV['app_baseLocalGeohistoryProjectUrl'] ?? '' ?>/<?= \Config\Services::request()->getLocale() ?>/<?= $row ?>
+<?= $_ENV['app_baseCanonicalProjectUrl'] ?? '' ?>/<?= \Config\Services::request()->getLocale() ?>/<?= $row ?>
 
 <?php }
     }